+++ src/collective/solr/behaviors.py
@@ -9,7 +9,7 @@ from zope.interface import provider
 
 @provider(IFormFieldProvider)
 class ISolrFields(model.Schema):
-    """ Social Media integration.
+    """ Additional fields to control Solr integration
     """
 
     directives.fieldset(
@@ -21,7 +21,7 @@ class ISolrFields(model.Schema):
         required=False,
         default=True,
         missing_value=True,
-        title=_('label_showinsearch', default=u"Show in search"),
+        title=_('label_showinsearch', default=u'Show in search'),
         description=_('help_showinsearch', default=''),
     )
 
@@ -29,13 +29,13 @@ class ISolrFields(model.Schema):
         required=False,
         default=[],
         missing_value=[],
-        title=_('label_searchwords', default=u"Search words"),
+        title=_('label_searchwords', default=u'Search words'),
         value_type=schema.TextLine(),
         description=_(
             'help_searchwords',
-            u"Specify words for which this item will show up "
-            u"as the first search result. Multiple words can be "
-            u"specified on new lines."
+            u'Specify words for which this item will show up '
+            u'as the first search result. Multiple words can be '
+            u'specified on new lines.'
         ),
     )