+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 6.0a2 (unreleased)
 ------------------
 
+- Implement reindexing the path indexes in solr. This means in solr path_string, path_parents and path_depth are updated on `obj.reindexObject(idxs=['path'])`.
+  [mathias.leimgruber]
+
 - Remove experimental.lazycatalog monkey patch and let
   SolrResponse inherit from ZCatalog Lazy class.
   [tomgross]