+++ CHANGES.rst
@@ -33,6 +33,9 @@ Changelog
 - Implement ``searchwords`` and ``showinsearch`` as Dexterity behavior.
   [tomgross]
 
+- Updated german translation
+  [tomgross]
+
 6.0a1 (2016-10-11)
 ------------------